Rename test-tool/helper/ to helper/, as other dirs are singular
