+++ src/job/test/JobFactory.spec.ts
@@ -1,4 +1,4 @@
-import AppTestHarness from 'test-tools/helpers/AppTestHarness';
+import AppTestHarness from 'test-tools/helper/AppTestHarness';
 import FakeEditorGateway from 'gateway/FakeEditorGateway';
 import JobFactory from 'job/JobFactory';
