+++ src/job/test/JobProvider.spec.ts
@@ -1,5 +1,5 @@
-import AppTestHarness from 'test-tools/helpers/AppTestHarness';
-import getContextStub from 'test-tools/helpers/getContextStub';
+import AppTestHarness from 'test-tools/helper/AppTestHarness';
+import getContextStub from 'test-tools/helper/getContextStub';
 import JobProviderFactory from 'job/JobProviderFactory';
 import JobFactory from 'job/JobFactory';
 import JobTreeItem from 'job/JobTreeItem';
